Fixes the regression test making it less strict.
Creating files with at the same time, with the same name and same
size but different content may trigger the SnapRAID check about
bad data. This happens in Mac OS X that likely has a not so precise
timestamp, with files created at the same time, even if using
nanosecond precision.